OAuth upgrade path dialog and handling (#598)
* OAuth upgrade path dialog and handling

* Use upgrade action to delegate the upgrade.

We thought that the problem was coming from auto-lock routing the app to the Welcome screen. We had a complex piece of code in the WelcomePresenter which was trying to manage the locking state and the upgrade state.

We've since worked out that there is not to do with auto-locking at all, just that the interactions between actions and observables and the iOS lifecycle mean reset is being called multiple times, and that the modal dialogue is being dismissed.

There seem to be two means of attack here: either avoid the double routing, or ensure that the dialog is shown each time the Welcome view appears. The former is the one most likely to succeed here, although we should ideally like the former.

The most concrete contribution is moving the versioning and detection of need for upgrade into the AppDelegate, where an LifecycleAction.upgrade action is launched. Multiple stores use this: the first is the DataStore which resets, causing the Welcome screen to be displayed. The second is the AccountStore which causes the _oldAccountPresence to be set to true (and thence to display the modal diag).

* use upgrade action, backfill specs, make sure reset works