Add basic Glean integration (#1129) (#1139)
* Add basic Glean integration (#1129)

This includes the minimum initialization for Glean, which will only enable and emit the Glean baseline ping.  This is tied to the "Send Usage Data" preference in order for users to be able to opt-out of telemetry collection.

This PR also adds and sets up a `GleanActionHandler` observer of the `TelemetryActions`, even though it's not doing anything with those actions yet.  This is in preparation for migration of existing telemetry to Glean but the actual collection of events and other metrics will be included in a subsequent PR.

Includes update to metrics.md to add Glean collection and references.

* Add comment clarifying `setUploadEnabled` is called before `initialize`

* Ensure the Glean observer of `recordUsageData` is used on main thread